APPS-50 Calculate values to be used. Use double everywhere.
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@65077 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/gui/view/table/SimpleTableValuesModel.java
@@ -20,7 +20,7 @@ class SimpleTableValuesModel implements TableValuesModel {
 
 	private List<TableValuesListener> listeners;
 	private List<Evaluatable> evaluatables;
-	private float[] values;
+	private double[] values;
 	private Kernel kernel;
 	private StringBuilder builder;
 
@@ -160,7 +160,7 @@ class SimpleTableValuesModel implements TableValuesModel {
 	 *
 	 * @param values x-values
 	 */
-	void setValues(float[] values) {
+	void setValues(double[] values) {
 		this.values = values;
 		ListIterator<String[]> iterator = columns.listIterator();
 		while (iterator.hasNext()) {
